move pragma injector and test into expected place
