+++ source/fab/tasks/c.py
@@ -6,14 +6,12 @@ C language handling classes.
 
 """
 import logging
-import re
 from collections import deque
-from typing import List, Pattern, Optional, Match
+from typing import List, Optional
 
 import clang.cindex  # type: ignore
 
 from fab.dep_tree import AnalysedFile
-from fab.tasks import TaskException
 from fab.util import log_or_dot, HashedFile
 
 logger = logging.getLogger(__name__)
@@ -134,34 +132,3 @@ class CAnalyser(object):
             analysed_file.add_symbol_dep(node.spelling)
 
 
-def CTextReaderPragmas(fpath):
-    """
-    Reads a C source file but when encountering an #include
-    preprocessor directive injects a special Fab-specific
-    #pragma which can be picked up later by the Analyser
-    after the preprocessing
-    """
-
-    _include_re: str = r'^\s*#include\s+(\S+)'
-    _include_pattern: Pattern = re.compile(_include_re)
-
-    for line in open(fpath, 'rt', encoding='utf-8'):
-        include_match: Optional[Match] = _include_pattern.match(line)
-        if include_match:
-            # For valid C the first character of the matched
-            # part of the group will indicate whether this is
-            # a system library include or a user include
-            include: str = include_match.group(1)
-            if include.startswith('<'):
-                yield '#pragma FAB SysIncludeStart\n'
-                yield line
-                yield '#pragma FAB SysIncludeEnd\n'
-            elif include.startswith(('"', "'")):
-                yield '#pragma FAB UsrIncludeStart\n'
-                yield line
-                yield '#pragma FAB UsrIncludeEnd\n'
-            else:
-                msg = 'Found badly formatted #include'
-                raise TaskException(msg)
-        else:
-            yield line
